% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_diamond.R
\name{dbinom_plot}
\alias{dbinom_plot}
\title{Plot binomial probability mass function}
\usage{
dbinom_plot(x, size, prob, show_table = FALSE)
}
\arguments{
\item{x}{Vector of integer values (number of successes)}

\item{size}{Number of trials}

\item{prob}{Probability of success}

\item{show_table}{Logical, whether to print probability table}
}
\description{
Plot binomial probability mass function
}
