% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_diamond.R
\name{fit_mean_diamond}
\alias{fit_mean_diamond}
\title{Draw Mean and Standard Deviation on a Plot}
\usage{
fit_mean_diamond(x, at = 1, col = "black", lwd = 2, length = 0.1)
}
\arguments{
\item{x}{Numeric vector.}

\item{at}{X-axis location to draw the line.}

\item{col}{Line color.}

\item{lwd}{Line width.}

\item{length}{Arrowhead length.}
}
\value{
No return value. Adds graphical elements to the current plot.
}
\description{
Adds an arrow representing the mean and standard deviation on an existing plot.
}
