% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_diamond.R
\name{glass.delta_1s}
\alias{glass.delta_1s}
\title{Glass's delta for single sample (bootstrap CI)}
\usage{
glass.delta_1s(x, mu, sd, R = 10000, conf.level = 0.95, seed = 123)
}
\arguments{
\item{x}{Numeric vector}

\item{mu}{Reference mean}

\item{sd}{Population SD}

\item{R}{Bootstrap reps}

\item{conf.level}{Confidence level}

\item{seed}{Random seed}
}
\value{
An object of class glassdelta
}
\description{
Glass's delta for single sample (bootstrap CI)
}
