% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_diamond.R
\name{mean_diamond}
\alias{mean_diamond}
\title{Compare Classical, Trimmed and Winsorized Means}
\usage{
mean_diamond(x, trim = 0.2, probs = c(0.05, 0.95), verbose = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{trim}{Trimming proportion for trimmed mean (default = 0.2).}

\item{probs}{Quantile cutoffs for winsorization (default = c(0.05, 0.95)).}

\item{verbose}{Logical. If \code{TRUE}, prints the resulting data frame.}
}
\value{
A data frame with mean, standard deviation, sample size, and method type.
}
\description{
Computes the mean and standard deviation using classical, trimmed, and winsorized methods.
}
