% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_diamond.R
\name{sdmean_diamond}
\alias{sdmean_diamond}
\title{Compute Mean and Standard Deviation}
\usage{
sdmean_diamond(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A named numeric vector with elements "Mean" and "SD".
}
\description{
Computes the mean and standard deviation of a numeric vector, ignoring NA values.
}
\examples{
sdmean_diamond(c(1, 2, 3, 4, NA))

}
